# api/swagger.yml
swagger: "2.0"
info:
  title: User manager
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
paths:
  /health/check:
    get:
      tags:
        - healthcheck
      responses:
        200:
          description: All services are available
          schema:
            type: object
            properties:
              resources:
                type: array
                items:
                  $ref: "#/definitions/HealthCheckResource"
        500:
          description: Part of services are unavailable
          schema:
            type: object
            properties:
              resources:
                type: array
                items:
                  $ref: "#/definitions/HealthCheckResource"

  /users:
    get:
      tags:
        - user
      operationId: getUsersByFilters
      parameters:
        - in: body
          name: body
          description: Filters for user searching
          schema:
            type: object
            required:
              - filters
            properties:
              filters:
                $ref: "#/definitions/Filters"
              limit:
                type: integer
                minimum: 1
                maximum: 20
                default: 10
                description: User count in response
              next:
                type: string
                description: Value is used for getting next user list. It is returned in response
                x-nullable: true

      responses:
        200:
          description: Users are filtered by criteria
          schema:
            $ref: "#/definitions/GetUserByFiltersResponse"
        422:
          description: Server couldn't handle request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal server error. Something went wrong
          schema:
            $ref: "#/definitions/Error"

    post:
      tags:
        - user
      operationId: createUser
      parameters:
        - in: body
          name: body
          description: User information for creating
          schema:
            $ref: "#/definitions/CreatingUser"
      responses:
        201:
          description: User was created successfully
          schema:
            $ref: "#/definitions/CreateUserResponse"
        422:
          description: Server couldn't handle request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal server error. Something went wrong
          schema:
            $ref: "#/definitions/Error"

    put:
      tags:
        - user
      operationId: updateUser
      parameters:
        - in: body
          name: body
          description: User information for updating
          schema:
            $ref: "#/definitions/UpdatingUser"
      responses:
        200:
          description: User was created successfully
          schema:
            $ref: "#/definitions/User"
        422:
          description: Server couldn't handle request
          schema:
            $ref: "#/definitions/Error"
        500:
          description: Internal server error. Something went wrong
          schema:
            $ref: "#/definitions/Error"

  /users/{userID}:
    delete:
      tags:
        - user
      operationId: deleteUser
      parameters:
        - $ref: "#/parameters/UserID"
      responses:
        204:
          description: User was created successfully
        500:
          description: Internal server error. Something went wrong
          schema:
            $ref: "#/definitions/Error"

parameters:
  UserID:
    in: path
    name: userID
    type: string
    format: uuid
    description: User identifier which was generated by service
    required: true

definitions:
  HealthCheckResource:
    type: object
    required:
      - resource
      - is_available
    properties:
      resource:
        type: string
        description: Checked resource
      is_available:
        type: boolean
        description: Show resource availability

  User:
    allOf:
      - $ref: "#/definitions/UserInfo"
      - type: object
        required:
          - id
          - password
        properties:
          id:
            type: string
            format: uuid
            description: User identifier which was generated by service
            x-nullable: false

  CreatingUser:
    allOf:
      - $ref: "#/definitions/UserInfo"
      - type: object
        required:
          - password
        properties:
          password:
            type: string
            #        pattern: ^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z]).{8,32}$
            description: User password. It must contains capital, small letters and digit
            minLength: 8
            maxLength: 256
            example: gD1wScAs
            x-nullable: false

  UpdatingUser:
    allOf:
      - $ref: "#/definitions/UserInfo"
      - type: object
        required:
          - id
          - password
        properties:
          id:
            type: string
            format: uuid
            description: User identifier which was generated by service
            x-nullable: false
          password:
            type: string
            #        pattern: ^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z]).{8,32}$
            description: User password. It must contains capital, small letters and digit
            minLength: 8
            maxLength: 256
            example: gD1wScAs
            x-nullable: false

  UserInfo:
    type: object
    required:
      - first_name
      - last_name
      - nickname
      - email
      - country_code
    properties:
      first_name:
        type: string
        description: User first name
        minLength: 1
        maxLength: 256
        example: Egor
        x-nullable: false
      last_name:
        type: string
        description: User last name
        minLength: 1
        maxLength: 256
        example: Shestakov
        x-nullable: false
      nickname:
        type: string
        description: User nickname
        maxLength: 32
        minLength: 2
        example: muzzzko
        x-nullable: false
      email:
        type: string
        pattern: (?:[a-z0-9!#$%&'*+\=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+\=?^_`{|}~-]+)*|"(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21\x23-\x5b\x5d-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])*")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\x01-\x08\x0b\x0c\x0e-\x1f\x21-\x5a\x53-\x7f]|\\[\x01-\x09\x0b\x0c\x0e-\x7f])+)\])
        description: User email
        maxLength: 256
        example: johnsmith@gmail.com
        x-nullable: false
      country_code:
        type: string
        minLength: 2
        maxLength: 2
        description: User country code
        example: 'UK'
        x-nullable: false

  Filters:
    type: object
    properties:
      country_code:
        type: string
        minLength: 2
        maxLength: 2
        description: User country code
        example: 'UK'
        x-nullable: true
    x-nullable: false

  Error:
    type: object
    properties:
      code:
        type: string
        description: Error code makes errors distinguish for client
        example: 1-1
      message:
        type: string
        description: Short message description
        example: object not found

  GetUserByFiltersResponse:
    type: object
    properties:
      users:
        type: array
        description: User list
        items:
          $ref: "#/definitions/User"
      next:
        type: string
        description: Value is used for getting next user list

  CreateUserResponse:
    type: object
    required:
      - id
    properties:
      id:
        type: string
        format: uuid
        description: User identifier which was generated by service
        x-nullable: false