# internal/generated/server/models/filters.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Filters filters
//
// swagger:model Filters
type Filters struct {

	// User country code
	// Example: UK
	// Max Length: 2
	// Min Length: 2
	CountryCode *string `json:"country_code,omitempty"`
}

// Validate validates this filters
func (m *Filters) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateCountryCode(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Filters) validateCountryCode(formats strfmt.Registry) error {
	if swag.IsZero(m.CountryCode) { // not required
		return nil
	}

	if err := validate.MinLength("country_code", "body", *m.CountryCode, 2); err != nil {
		return err
	}

	if err := validate.MaxLength("country_code", "body", *m.CountryCode, 2); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this filters based on context it is used
func (m *Filters) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Filters) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Filters) UnmarshalBinary(b []byte) error {
	var res Filters
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}