# internal/generated/server/models/updating_user.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// UpdatingUser updating user
//
// swagger:model UpdatingUser
type UpdatingUser struct {
	UserInfo

	// User identifier which was generated by service
	// Required: true
	// Format: uuid
	ID strfmt.UUID `json:"id"`

	// User password. It must contains capital, small letters and digit
	// Example: gD1wScAs
	// Required: true
	// Max Length: 256
	// Min Length: 8
	Password string `json:"password"`
}

// UnmarshalJSON unmarshals this object from a JSON structure
func (m *UpdatingUser) UnmarshalJSON(raw []byte) error {
	// AO0
	var aO0 UserInfo
	if err := swag.ReadJSON(raw, &aO0); err != nil {
		return err
	}
	m.UserInfo = aO0

	// AO1
	var dataAO1 struct {
		ID strfmt.UUID `json:"id"`

		Password string `json:"password"`
	}
	if err := swag.ReadJSON(raw, &dataAO1); err != nil {
		return err
	}

	m.ID = dataAO1.ID

	m.Password = dataAO1.Password

	return nil
}

// MarshalJSON marshals this object to a JSON structure
func (m UpdatingUser) MarshalJSON() ([]byte, error) {
	_parts := make([][]byte, 0, 2)

	aO0, err := swag.WriteJSON(m.UserInfo)
	if err != nil {
		return nil, err
	}
	_parts = append(_parts, aO0)
	var dataAO1 struct {
		ID strfmt.UUID `json:"id"`

		Password string `json:"password"`
	}

	dataAO1.ID = m.ID

	dataAO1.Password = m.Password

	jsonDataAO1, errAO1 := swag.WriteJSON(dataAO1)
	if errAO1 != nil {
		return nil, errAO1
	}
	_parts = append(_parts, jsonDataAO1)
	return swag.ConcatJSON(_parts...), nil
}

// Validate validates this updating user
func (m *UpdatingUser) Validate(formats strfmt.Registry) error {
	var res []error

	// validation for a type composition with UserInfo
	if err := m.UserInfo.Validate(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateID(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validatePassword(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *UpdatingUser) validateID(formats strfmt.Registry) error {

	if err := validate.Required("id", "body", strfmt.UUID(m.ID)); err != nil {
		return err
	}

	if err := validate.FormatOf("id", "body", "uuid", m.ID.String(), formats); err != nil {
		return err
	}

	return nil
}

func (m *UpdatingUser) validatePassword(formats strfmt.Registry) error {

	if err := validate.RequiredString("password", "body", m.Password); err != nil {
		return err
	}

	if err := validate.MinLength("password", "body", m.Password, 8); err != nil {
		return err
	}

	if err := validate.MaxLength("password", "body", m.Password, 256); err != nil {
		return err
	}

	return nil
}

// ContextValidate validate this updating user based on the context it is used
func (m *UpdatingUser) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	var res []error

	// validation for a type composition with UserInfo
	if err := m.UserInfo.ContextValidate(ctx, formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

// MarshalBinary interface implementation
func (m *UpdatingUser) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *UpdatingUser) UnmarshalBinary(b []byte) error {
	var res UpdatingUser
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}